Fix the binary-arch rule
Currently the binary-arch rule is broken (when doing a pure
arch-dependent build), as the dh_pysupport called by the binary-common
rule is only provided by a build-depends-indep, and not a build-depends.

This patch adds a makefile ifeq conditional, such that we only call
dh_pysupport when in arch-independent build mode.