Partially re-enable parallel builds
Commit 0c3cc46ed disabled parallel builds with this message:

“Disable parallel builds, because libtool is made of fail (if
binary-arch and build-indep run concurrently, we relink a library
while it's being used; that doesn't work so well)”.

So the problem is that MAKEFLAGS is exported to even the make children
running the debian rules, leading to concurrent running of multiple
rules. But if we don't export MAKEFLAGS globally, just pass
appropriate ‘-j’ options to the expensive rules, we solve the intended
goal (speedy compilation on beefy machines) without breaking the
builds.